Thru BL/PL test

+++ DTB.DVDCentral/DTB.DVDCentral.BL.Models/Movie.cs
@@ -11,13 +11,11 @@ namespace DTB.DVDCentral.BL.Models
         public int Id { get; set; }
         public string Title { get; set; }
         public string Description { get; set; }
-        public double Cost { get; set; }
+        public decimal Cost { get; set; }
         public int RatingId { get; set; }
         public int FormatId { get; set; }
         public int DirectorId { get; set; }
         public int InStkQty { get; set; }
         public string ImagePath { get; set; }
-
-
     }
 }
